backend/requirements.txt
fastapi==0.116.1
passlib==1.7.4
pydantic==2.11.7
python-dotenv==1.1.1
python_jose==3.5.0
SQLAlchemy==2.0.43
uvicorn==0.15.0